+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "xamlstyler" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Added
+- Respect indentation options of VSCode IDE
+
 ## [0.0.5]
 
 ### Added